+++ LifeSim/rand.h
@@ -7,4 +7,4 @@
 
 float GetDistribution(std::uniform_real_distribution<float> distro);
 int GetDistribution(std::uniform_int_distribution<int> distro);
-int GetDistribution(std::normal_distribution<int> distro);
+float GetDistribution(std::normal_distribution<float> distro);
